refactor: type declare file
